+++ apis/cds.d.ts
@@ -6,6 +6,7 @@ export * from './services'
 export * from './events'
 export * from './utils'
 export * from './cqn'
+export * from './global'
 export { log, debug } from './log'
 export { test } from './test'
 
@@ -22,23 +23,3 @@ export { QLExtensions } from './ql' // cds-ql.ts test tries to import this from 
 import { Service } from './services'
 declare const delete_: Service['delete']
 export { delete_ as delete }
-
-declare global {
-  // these provide the functionality from SELECT, INSERT, etc in the global facade
-  const SELECT: ql.QL<any>['SELECT']
-  const INSERT: ql.QL<any>['INSERT']
-  const UPSERT: ql.QL<any>['UPSERT']
-  const UPDATE: ql.QL<any>['UPDATE']
-  const DELETE: ql.QL<any>['DELETE']
-  const CREATE: ql.QL<any>['CREATE']
-  const DROP: ql.QL<any>['DROP']
-
-	// and these allow us to use them as type too, i.e. `const q: SELECT<Book> = ...`
-	type SELECT<T> = ql.SELECT<T>
-	type INSERT<T> = ql.INSERT<T>
-	type UPSERT<T> = ql.UPSERT<T>
-	type UPDATE<T> = ql.UPDATE<T>
-	type DELETE<T> = ql.DELETE<T>
-	type CREATE<T> = ql.CREATE<T>
-	type DROP<T> = ql.DROP<T>
-}
